pip-511: move views to pages folder
