--- src/main/routes/home.ts
@@ -1,9 +1,0 @@
-import { Application } from 'express';
-
-export default function(app: Application): void {
-
-  app.get('/', (req, res) => {
-    res.render('home');
-  });
-
-}
