+++ src/main/routes/pages/health.ts
@@ -1,0 +1,14 @@
+import { Application } from 'express';
+
+const healthcheck = require('@hmcts/nodejs-healthcheck');
+
+export default function(app: Application): void {
+  const healthCheckConfig = {
+    checks: {
+      // TODO: replace this sample check with proper checks for your application
+      sampleCheck: healthcheck.raw(() => healthcheck.up()),
+    },
+  };
+
+  healthcheck.addTo(app, healthCheckConfig);
+}
