+++ src/main/routes/pages/home.ts
@@ -1,0 +1,9 @@
+import { Application } from 'express';
+
+export default function(app: Application): void {
+
+  app.get('/', (req, res) => {
+    res.render('home');
+  });
+
+}
